backend/requirements.txt
trimesh>=3.20.0
numpy>=1.24.0
scipy>=1.10.0
pymongo[srv]
python-dotenv
dnspython
fastapi
uvicorn
python-multipart
pydantic
